feat: refactor config and fix windows
