+++ cmd/commands/start/start.go
@@ -6,11 +6,10 @@ package start
 import (
 	"observe/agent/cmd"
 	observeotel "observe/agent/cmd/collector"
-	"observe/agent/cmd/connections"
+	"observe/agent/cmd/config"
 	"os"
 
 	"github.com/spf13/cobra"
-	"github.com/spf13/viper"
 )
 
 var startCmd = &cobra.Command{
@@ -20,31 +19,19 @@ var startCmd = &cobra.Command{
 This command reads in the local config and env vars and starts the 
 collector on the current host.`,
 	RunE: func(cmd *cobra.Command, args []string) error {
-		// Handle base OTEL config
-		err := observeotel.SetEnvVars()
+		// Set Env Vars from config
+		err := config.SetEnvVars()
 		if err != nil {
 			return err
 		}
-		// Initialize config file paths with base config
-		configFilePaths := []string{observeotel.BaseOtelCollectorConfigFilePath}
-		// Get additional config paths based on connection configs
-		if viper.IsSet(connections.HostMonitoringConnectionType.Name) {
-			configFilePaths = append(configFilePaths, connections.HostMonitoringConnectionType.GetConfigFilePaths()...)
+		//
+		configFilePaths, overridePath, err := config.GetAllOtelConfigFilePaths()
+		if err != nil {
+			return err
 		}
-		// Generate override file and include path if overrides provided
-		var overridePath string
-		if viper.IsSet("otel_config_overrides") {
-			overridePath, err = observeotel.GetOverrideConfigFile(viper.Sub("otel_config_overrides"))
-			if err != nil {
-				return err
-			}
-			configFilePaths = append(configFilePaths, overridePath)
+		if overridePath != "" {
+			defer os.Remove(overridePath)
 		}
-		defer func() {
-			if viper.IsSet("otel_config_overrides") {
-				os.Remove(overridePath)
-			}
-		}()
 		// Generate collector settings with all config files
 		colSettings := observeotel.GenerateCollectorSettings(configFilePaths)
 		otelCmd := observeotel.GetOtelCollectorCommand(colSettings)
