+++ main_windows.go
@@ -6,13 +6,14 @@ import (
 	"errors"
 	"fmt"
 	"log"
-	observeotel "observe/agent/cmd/connections/otelcollector"
+	observeotel "observe/agent/cmd/collector"
 	"os"
 
 	"go.opentelemetry.io/collector/otelcol"
 	"golang.org/x/sys/windows"
 
 	"observe/agent/cmd"
+	"observe/agent/cmd/config"
 
 	"golang.org/x/sys/windows/svc"
 )
@@ -29,10 +30,21 @@ func run() error {
 		}
 		cmd.CfgFile = os.Args[1]
 		cmd.InitConfig()
-		observeotel.SetEnvVars()
-
-		params := observeotel.GenerateCollectorSettings()
-		if err := svc.Run("", otelcol.NewSvcHandler(*params)); err != nil {
+		// Set Env Vars from config
+		err := config.SetEnvVars()
+		if err != nil {
+			return err
+		}
+		//
+		configFilePaths, overridePath, err := config.GetAllOtelConfigFilePaths()
+		if err != nil {
+			return err
+		}
+		if overridePath != "" {
+			defer os.Remove(overridePath)
+		}
+		colSettings := observeotel.GenerateCollectorSettings(configFilePaths)
+		if err := svc.Run("", otelcol.NewSvcHandler(*colSettings)); err != nil {
 			if errors.Is(err, windows.ERROR_FAILED_SERVICE_CONTROLLER_CONNECT) {
 				// Per https://learn.microsoft.com/en-us/windows/win32/api/winsvc/nf-winsvc-startservicectrldispatchera#return-value
 				// this means that the process is not running as a service, so run interactively.
